-set easy password -moved applicationDBcontext to DAL folder -enabled migrations
