+++ ConferenceFormSubmittal/App_Start/IdentityConfig.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin;
 using Microsoft.Owin.Security;
 using ConferenceFormSubmittal.Models;
+using ConferenceFormSubmittal.DAL;
 
 namespace ConferenceFormSubmittal
 {
@@ -54,10 +55,10 @@ namespace ConferenceFormSubmittal
             manager.PasswordValidator = new PasswordValidator
             {
                 RequiredLength = 6,
-                RequireNonLetterOrDigit = true,
-                RequireDigit = true,
-                RequireLowercase = true,
-                RequireUppercase = true,
+                RequireNonLetterOrDigit = false,
+                RequireDigit = false,
+                RequireLowercase = false,
+                RequireUppercase = false,
             };
 
             // Configure user lockout defaults
