+++ ConferenceFormSubmittal/App_Start/Startup.Auth.cs
@@ -6,6 +6,7 @@ using Microsoft.Owin.Security.Cookies;
 using Microsoft.Owin.Security.Google;
 using Owin;
 using ConferenceFormSubmittal.Models;
+using ConferenceFormSubmittal.DAL;
 
 namespace ConferenceFormSubmittal
 {
