+++ ConferenceFormSubmittal/ConferenceFormSubmittal.csproj
@@ -205,6 +205,7 @@
     <Compile Include="Controllers\ManageController.cs" />
     <Compile Include="Controllers\MileagesController.cs" />
     <Compile Include="Controllers\SitesController.cs" />
+    <Compile Include="DAL\ApplicationDBContext.cs" />
     <Compile Include="DAL\CFSEntities.cs" />
     <Compile Include="DAL\CFSMigrations\201802112120261_InitialCreate.cs" />
     <Compile Include="DAL\CFSMigrations\201802112120261_InitialCreate.Designer.cs">
@@ -227,6 +228,7 @@
       <DependentUpon>201803201059446_FileUploads.cs</DependentUpon>
     </Compile>
     <Compile Include="DAL\CFSMigrations\Configuration.cs" />
+    <Compile Include="DAL\SecurityMigrations\Configuration.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
