+++ ConferenceFormSubmittal/DAL/SecurityMigrations/Configuration.cs
@@ -1,0 +1,21 @@
+namespace ConferenceFormSubmittal.DAL.SecurityMigrations
+{
+    using System;
+    using System.Data.Entity;
+    using System.Data.Entity.Migrations;
+    using System.Linq;
+
+    internal sealed class Configuration : DbMigrationsConfiguration<ConferenceFormSubmittal.DAL.ApplicationDbContext>
+    {
+        public Configuration()
+        {
+            AutomaticMigrationsEnabled = false;
+            MigrationsDirectory = @"DAL\SecurityMigrations";
+        }
+
+        protected override void Seed(ConferenceFormSubmittal.DAL.ApplicationDbContext context)
+        {
+
+        }
+    }
+}
