+++ ConferenceFormSubmittal/Models/IdentityModels.cs
@@ -16,18 +16,10 @@ namespace ConferenceFormSubmittal.Models
             // Add custom user claims here
             return userIdentity;
         }
+
+        //Add special fields here
+        //public int FavIceCream { get; set; }
     }
 
-    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
-    {
-        public ApplicationDbContext()
-            : base("DefaultConnection", throwIfV1Schema: false)
-        {
-        }
-
-        public static ApplicationDbContext Create()
-        {
-            return new ApplicationDbContext();
-        }
-    }
+   
 }
